requirements.txt
azure-core==1.31.0
azure-eventhub==5.12.2
azure-functions==1.21.3
azure-identity==1.19.0
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
hl7apy==1.3.5
hl7parser==0.7.4
idna==3.10
isodate==0.7.2
msal==1.31.0
msal-extensions==1.2.0
numpy==2.1.2
pandas==2.2.3
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
